+++ core/src/com/mygdx/game/screens/SettingsScreen.java
@@ -119,7 +119,7 @@ public class SettingsScreen implements Screen {
 
                 try{
                     File f = new File(savePath.getText());
-                    if (!f.createNewFile()) {
+                    if (!IO_course_module.isDefaultCourseName(savePath.getText()) && !f.createNewFile()) {
                         f.delete();
                         f.createNewFile();
                     }
